requirements.txt
torch==2.7.1+cu126
torchvision==0.22.1+cu126
fastapi==0.115.12
pydantic==2.11.5
cerebrium==1.40.4
uvicorn==0.29.0
pillow==11.0.0
onnx==1.18.0
onnxruntime==1.22.0